refactor : 파일 이름 역할에 맞게 재설정 (#63)
