+++ src/solchat/core/embedding_model.py
