+++ src/solchat/core/exception_handlers.py
