+++ src/solchat/core/llm_key_manager.py
