+++ src/solchat/core/utils/stream_utils.py
