+++ src/solchat/core/v1/prompt_templates.py
