+++ src/solchat/core/v2/chat_prompt_templates.py
