+++ src/solchat/core/v2/prompt_templates_v2.py
