+++ src/solchat/docker-entrypoint.sh
