+++ src/solchat/schemas/v1/solution_schema.py
