+++ src/solchat/schemas/v2/interview_schema.py
