+++ src/solchat/schemas/v2/solution_schema_v2.py
